add code and data for condition model 3
model 3 has the interaction between stream and smb. Run condition  model 3 stan runs the model, condition_model_3_out.rds is the output that comes from running the model that can then be loaded to make the plots, and condition_model_3_TEST.stan is the stan code. Plots can be found in run condition model 3 stan.R and can be run if you load the data set on line 45.